Silenced some warnings about implicit casting

+++ include/processes/swarm_process.hpp
@@ -9,7 +9,7 @@ namespace SwarmSystem
 {
 void update(
     entt::registry &,
-    const float);
+    const double);
 };
 
 #endif
